backend/pyproject.toml
[tool.poetry]
name = "kagomma"
version = "0.0.1"
description = ""
authors = ["hitokun-s <jgpuauno@gmail.com>"]
readme = "README.md"
#packages = [{include = "kagomma", from="src"}]

[tool.poetry.dependencies]
#python = "^3.9"
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
#faiss-cpu = "^1.7.3"
#langchain = "^0.0.135"
openai = "^0.27.4"
#tiktoken = "^0.3.3"
#google-search-results = "^2.4.2"
#chromadb = "^0.3.21"
#jupyter = "^1.0.0"
#wikipedia = "^1.4.0"
python-dotenv = "^1.0.0"
#transformers = "^4.27.4"
#streamlit-chat = "^0.0.2.2"
#bokeh = "2.4.2"
#streamlit-bokeh-events = "^0.1.2"
#gtts = "^2.3.1"
google-api-python-client = "^2.85.0"
#streamlit = "^1.22.0"
#streamlit-pills = "^0.3.0"
#streamlit-extras = "^0.2.7"
joblib = "^1.2.0"
flask = "^2.3.2"
flask-cors = "^3.0.10"
#flask-jwt-extended = "^4.4.4"
#flask-login = "^0.6.2"
#flask-redis = "^0.4.0"
gunicorn = "^20.1.0"
rich = "^13.3.5"
flask-limiter = "^3.3.1"
flask-session = "^0.5.0"
redis = "^4.5.5"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
mypy = "^1.3.0"
pylint = "^2.17.4"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
